Move smoke tests to the main integration tests file
