+++ tests/integration/specs/analyze/analyze-dependency-risks.test.ts
@@ -24,15 +24,220 @@
 // list of dependency risks. Once the real scanner is wired, the happy-path
 // assertions will be expanded.
 
-import { afterEach, beforeEach, describe, expect, it } from 'bun:test';
+import { afterEach, beforeEach, describe, expect, it, spyOn } from 'bun:test';
 
 import { buildLocalBinaryName } from '../../../../src/cli/commands/_common/install/sca-scanner.js';
+import { DefaultScaScannerSpawner } from '../../../../src/cli/commands/analyze/dependency-risk-helpers/default-sca-scanner-spawner.js';
+import type { AnalyzeProjectResponse } from '../../../../src/cli/commands/analyze/dependency-risk-helpers/sca-scanner.js';
+import { analyzeDependencyRisks } from '../../../../src/cli/commands/analyze/dependency-risks.js';
+import type { ResolvedAuth } from '../../../../src/lib/auth-resolver.js';
 import { detectPlatform } from '../../../../src/lib/platform-detector.js';
+import { clearMockUiCalls, getMockUiCalls, setMockUi } from '../../../../src/ui/index.js';
 import { TestHarness } from '../../harness';
 
 const VALID_TOKEN = 'integration-test-token';
 const TEST_ORG = 'my-org';
 const SCA_SCANNER_FAILURE_PREFIX = 'Dependency risk analysis error: sca-scanner exited with code';
+
+const SCAN_RESULT_STUB: AnalyzeProjectResponse = {
+  releases: [
+    {
+      key: 'lodash:4.17.21',
+      packageUrl: 'pkg:npm/lodash@4.17.21',
+      packageManager: 'npm',
+      packageName: 'lodash',
+      version: '4.17.21',
+      licenseExpression: 'MIT',
+      known: true,
+      knownPackage: true,
+      newlyIntroduced: false,
+      dependencyFilePaths: ['package-lock.json'],
+      dependencyChains: [['pkg:npm/lodash@4.17.21']],
+      issues: [
+        {
+          key: 'issue-1',
+          severity: 'HIGH',
+          showIncreasedSeverityWarning: null,
+          type: 'VULNERABILITY',
+          quality: 'SECURITY',
+          status: 'ACCEPT',
+          vulnerabilityId: 'CVE-2024-0001',
+          cweIds: null,
+          cvssScore: '7.5',
+          spdxLicenseId: null,
+          versionOptions: null,
+        },
+      ],
+    },
+  ],
+  parsedFiles: ['package-lock.json'],
+  errors: [],
+};
+
+const EXPECTED_JSON = `{
+  "project": "my-project",
+  "packages": [
+    {
+      "package": "pkg:npm/lodash@4.17.21",
+      "newlyIntroduced": false,
+      "riskCount": 1,
+      "filePaths": [
+        "package-lock.json"
+      ],
+      "chains": [
+        {
+          "parentPackage": null,
+          "chains": [
+            [
+              "pkg:npm/lodash@4.17.21"
+            ]
+          ]
+        }
+      ],
+      "groups": [
+        {
+          "type": "VULNERABILITY",
+          "selectedRisks": [
+            {
+              "severity": "HIGH",
+              "status": "ACCEPT",
+              "cvssScore": "7.5",
+              "vulnerabilityId": "CVE-2024-0001",
+              "partialFixes": []
+            }
+          ],
+          "recommendation": {
+            "action": "NO_FIX_AVAILABLE",
+            "fixVersions": []
+          },
+          "totalKnownRisksCount": 1
+        }
+      ]
+    }
+  ],
+  "errors": [],
+  "summary": {
+    "packagesScanned": 1,
+    "totalRisks": 1,
+    "byType": {
+      "MALWARE": {
+        "BLOCKER": 0,
+        "HIGH": 0,
+        "MEDIUM": 0,
+        "LOW": 0,
+        "INFO": 0
+      },
+      "PROHIBITED_LICENSE": {
+        "BLOCKER": 0,
+        "HIGH": 0,
+        "MEDIUM": 0,
+        "LOW": 0,
+        "INFO": 0
+      },
+      "VULNERABILITY": {
+        "BLOCKER": 0,
+        "HIGH": 1,
+        "MEDIUM": 0,
+        "LOW": 0,
+        "INFO": 0
+      }
+    },
+    "packages": [
+      {
+        "package": "pkg:npm/lodash@4.17.21",
+        "riskCount": 1,
+        "highestSeverity": "HIGH",
+        "recommendations": {
+          "VULNERABILITY": {
+            "action": "NO_FIX_AVAILABLE",
+            "fixVersions": []
+          }
+        }
+      }
+    ],
+    "filter": {
+      "effectiveStatuses": [
+        "NEW",
+        "OPEN",
+        "CONFIRM",
+        "ACCEPT",
+        "SAFE",
+        "FIXED"
+      ],
+      "discardedStatuses": []
+    }
+  }
+}`;
+
+const EXPECTED_TOON = `project: my-project
+packages[1]:
+  - package: "pkg:npm/lodash@4.17.21"
+    newlyIntroduced: false
+    riskCount: 1
+    filePaths[1]: package-lock.json
+    chains[1]{parentPackage,totalRoutes}:
+      direct,1
+    groups[1]:
+      - type: VULNERABILITY
+        selectedRisks[1]{severity,status,cvssScore,vulnerabilityId}:
+          HIGH,ACCEPT,"7.5",CVE-2024-0001
+        recommendation:
+          action: NO_FIX_AVAILABLE
+          fixVersions[0]:
+        totalKnownRisksCount: 1
+errors[0]:
+summary:
+  packagesScanned: 1
+  totalRisks: 1
+  byType:
+    MALWARE:
+      BLOCKER: 0
+      HIGH: 0
+      MEDIUM: 0
+      LOW: 0
+      INFO: 0
+    PROHIBITED_LICENSE:
+      BLOCKER: 0
+      HIGH: 0
+      MEDIUM: 0
+      LOW: 0
+      INFO: 0
+    VULNERABILITY:
+      BLOCKER: 0
+      HIGH: 1
+      MEDIUM: 0
+      LOW: 0
+      INFO: 0
+  packages[1]:
+    - package: "pkg:npm/lodash@4.17.21"
+      riskCount: 1
+      highestSeverity: HIGH
+      recommendations:
+        VULNERABILITY:
+          action: NO_FIX_AVAILABLE
+          fixVersions[0]:
+  filter:
+    effectiveStatuses[6]: NEW,OPEN,CONFIRM,ACCEPT,SAFE,FIXED
+    discardedStatuses[0]:`;
+
+const EXPECTED_TABLE = `── lodash@4.17.21 (1 risk) ─────────────────────────────────────────────────────
+in: package-lock.json
+  direct
+
+  HIGH      ACCEPT   CVSS 7.5 CVE-2024-0001 → no fix available
+  No recommended version without known vulnerabilities
+
+════════════════════════════════════════════════════════════════════════════════
+
+Summary: 1 dependencies checked, 1 risks found
+Filtering by: new, open, confirm, accept, safe, fixed
+  MALWARE             BLOCKER ✓   0    HIGH ✓   0    MEDIUM ✓   0    LOW ✓   0    INFO ✓   0
+  PROHIBITED_LICENSE  BLOCKER ✓   0    HIGH ✓   0    MEDIUM ✓   0    LOW ✓   0    INFO ✓   0
+  VULNERABILITY       BLOCKER ✓   0    HIGH ✗   1    MEDIUM ✓   0    LOW ✓   0    INFO ✓   0
+
+Recommendations:
+  lodash@4.17.21 (1 risk, highest severity HIGH)
+    No recommended version without known vulnerabilities`;
 
 describe('analyze dependency-risks', () => {
   let harness: TestHarness;
@@ -269,5 +474,72 @@ describe('analyze dependency-risks', () => {
       'Software Composition Analysis is not available for the current connection.',
     );
     expect(server.getRecordedRequests().some((r) => r.path === '/api/system/status')).toBe(false);
+  });
+
+  describe('output format', () => {
+    let spawnSpy: ReturnType<typeof spyOn>;
+    let auth: ResolvedAuth;
+
+    beforeEach(async () => {
+      const server = await harness
+        .newFakeServer()
+        .withAuthToken(VALID_TOKEN)
+        .withScaEnabled(true)
+        .withProject('my-project')
+        .withProjectSettings('my-project', [])
+        .start();
+      auth = {
+        token: VALID_TOKEN,
+        serverUrl: server.baseUrl(),
+        orgKey: TEST_ORG,
+        connectionType: 'cloud',
+      };
+
+      setMockUi(true);
+      spawnSpy = spyOn(DefaultScaScannerSpawner.prototype, 'spawn').mockResolvedValue({
+        exitCode: 0,
+        stdout: JSON.stringify(SCAN_RESULT_STUB),
+        stderr: '',
+      });
+    });
+
+    afterEach(() => {
+      spawnSpy.mockRestore();
+      setMockUi(false);
+      clearMockUiCalls();
+    });
+
+    function getPrinted(): string {
+      const call = getMockUiCalls().find((c) => c.method === 'print');
+      if (!call) throw new Error('expected print() to be called');
+      return call.args[0] as string;
+    }
+
+    it('serializes as JSON for --format json', async () => {
+      await analyzeDependencyRisks(
+        { project: 'my-project', format: 'json', statuses: 'all' },
+        auth,
+      );
+
+      expect(getPrinted()).toBe(EXPECTED_JSON);
+    });
+
+    it('serializes as TOON for --format toon', async () => {
+      await analyzeDependencyRisks(
+        { project: 'my-project', format: 'toon', statuses: 'all' },
+        auth,
+      );
+
+      expect(getPrinted()).toBe(EXPECTED_TOON);
+    });
+
+    it('serializes as a table for the default format', async () => {
+      await analyzeDependencyRisks(
+        { project: 'my-project', format: 'table', statuses: 'all' },
+        auth,
+      );
+
+      expect(getPrinted()).toBe(EXPECTED_TABLE);
+    });
   });
 });
